Move helper functions to test fixtures
These were only called from test code so can be removed from the public API
