+++ detekt-api/src/testFixtures/kotlin/io/gitlab/arturbosch/detekt/test/TestFactory.kt
@@ -10,6 +10,7 @@ import io.gitlab.arturbosch.detekt.api.Severity
 import io.gitlab.arturbosch.detekt.api.SourceLocation
 import io.gitlab.arturbosch.detekt.api.TextLocation
 import org.jetbrains.kotlin.psi.KtElement
+import java.nio.file.Path
 import kotlin.io.path.Path
 
 fun createIssue(
@@ -63,7 +64,7 @@ fun createIssueForRelativePath(
             source = SourceLocation(1, 1),
             endSource = SourceLocation(1, 1),
             text = TextLocation(0, 0),
-            filePath = FilePath.fromRelative(Path(basePath), Path(relativePath))
+            filePath = fromRelative(Path(basePath), Path(relativePath))
         ),
         ktElement = null
     ),
@@ -91,8 +92,8 @@ fun createLocation(
     source = SourceLocation(position.first, position.second),
     endSource = SourceLocation(endPosition.first, endPosition.second),
     text = TextLocation(text.first, text.last),
-    filePath = basePath?.let { FilePath.fromRelative(Path(it), Path(path)) }
-        ?: FilePath.fromAbsolute(Path(path)),
+    filePath = basePath?.let { fromRelative(Path(it), Path(path)) }
+        ?: fromAbsolute(Path(path)),
 )
 
 private data class IssueImpl(
@@ -109,3 +110,10 @@ private data class IssueImpl(
         override val description: String,
     ) : Issue.RuleInfo
 }
+
+fun fromAbsolute(path: Path) = FilePath(absolutePath = path.normalize())
+fun fromRelative(basePath: Path, relativePath: Path) = FilePath(
+    absolutePath = basePath.resolve(relativePath).normalize(),
+    basePath = basePath.normalize(),
+    relativePath = relativePath
+)
