+++ detekt-psi-utils/api/detekt-psi-utils.api
@@ -4,18 +4,12 @@ public final class io/github/detekt/psi/AnnotationExcluder {
 }
 
 public final class io/github/detekt/psi/FilePath {
-	public static final field Companion Lio/github/detekt/psi/FilePath$Companion;
 	public fun <init> (Ljava/nio/file/Path;Ljava/nio/file/Path;Ljava/nio/file/Path;)V
 	public synthetic fun <init> (Ljava/nio/file/Path;Ljava/nio/file/Path;Ljava/nio/file/Path;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public final fun getAbsolutePath ()Ljava/nio/file/Path;
 	public final fun getBasePath ()Ljava/nio/file/Path;
 	public final fun getRelativePath ()Ljava/nio/file/Path;
 	public fun toString ()Ljava/lang/String;
-}
-
-public final class io/github/detekt/psi/FilePath$Companion {
-	public final fun fromAbsolute (Ljava/nio/file/Path;)Lio/github/detekt/psi/FilePath;
-	public final fun fromRelative (Ljava/nio/file/Path;Ljava/nio/file/Path;)Lio/github/detekt/psi/FilePath;
 }
 
 public abstract class io/github/detekt/psi/FunctionMatcher {
